refactor : 최대 indent 3 넘지 않도록 함수 분리 및 폴더 구조 변경
