+++ docs/README.md
@@ -33,6 +33,6 @@
 
 ### Todo
 
-- [ ] 자동차 이름 길이, 쉼표 기준 등 input test
-- [ ] App파일 indent 3 이상 넘지 않도록 함수화
-- [ ] 상수들 변수화 하기
+- [x] 자동차 이름 길이, 쉼표 기준 등 input test 파일 작성
+- [x] App파일 indent 3 이상 넘지 않도록 함수화
+- [x] 상수들 변수화 하기
